content/canvas/test/test_type.exists.html

<!DOCTYPE HTML>
<title>Canvas test: type.exists</title>
<!-- Testing: HTMLCanvasElement is a property of window -->
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

ok(window.HTMLCanvasElement, "window.HTMLCanvasElement");

SimpleTest.finish();

});
</script>
